+++ test/test.js
@@ -26,6 +26,16 @@ describe('Test ARD Package', () => {
 
 // Test Helpers Package
 describe('Test Helpers Package', () => {
+	describe('Test getJsonKeys', () => {
+		it("getJsonKeys({ hello: 'world', foo: 'bar' }) = ['hello', 'foo']", () => {
+			const test = { hello: 'world', foo: 'bar' }
+			const result = ['hello', 'foo']
+			const testResult = helpers.getJsonKeys(test)
+			expect(testResult[0]).to.equal(result[0])
+			expect(testResult[1]).to.equal(result[1])
+		})
+	})
+
 	describe('Test sleep', () => {
 		it('sleep(1e3) will sleep 1s', async () => {
 			const time = 1e3
@@ -155,30 +165,30 @@ describe('Test Numbers Package', () => {
 // Test Strings Package
 describe('Test Strings Package', () => {
 	describe('Test capitalize', () => {
-		it('capitalize("a") = A', () => {
+		it("capitalize('a') = 'A'", () => {
 			expect(strings.capitalize('a')).to.equal('A')
 		})
-		it('capitalize("apple") = Apple', () => {
+		it("capitalize('apple') = 'Apple'", () => {
 			expect(strings.capitalize('apple')).to.equal('Apple')
 		})
 	})
 
 	describe('Test getObjectLength', () => {
-		it('getObjectLength({ hello: "world" }) = 1', () => {
+		it("getObjectLength({ hello: 'world' }) = 1", () => {
 			expect(strings.getObjectLength({ hello: 'world' })).to.equal(1)
 		})
 
-		it('getObjectLength({ hello: "world", foo: "bar" }) = 2', () => {
+		it("getObjectLength({ hello: 'world', foo: 'bar' }) = 2", () => {
 			expect(strings.getObjectLength({ hello: 'world', foo: 'bar' })).to.equal(2)
 		})
 	})
 
 	describe('Test isArray', () => {
-		it('isArray(["hello world"]) = true', () => {
+		it("isArray(['hello world']) = true", () => {
 			expect(strings.isArray(['hello world'])).to.equal(true)
 		})
 
-		it('isArray({ hello: "world" }) = false', () => {
+		it("isArray({ hello: 'world' }) = false", () => {
 			expect(strings.isArray({ hello: 'world' })).to.equal(false)
 		})
 	})
@@ -188,7 +198,7 @@ describe('Test Strings Package', () => {
 			expect(strings.isEmptyArray([])).to.equal(true)
 		})
 
-		it('isEmptyArray(["hello world"]) = false', () => {
+		it("isEmptyArray(['hello world']) = false", () => {
 			expect(strings.isEmptyArray(['hello world'])).to.equal(false)
 		})
 	})
@@ -198,27 +208,27 @@ describe('Test Strings Package', () => {
 			expect(strings.isEmptyObject({})).to.equal(true)
 		})
 
-		it('isEmptyObject({ hello: "world" }) = false', () => {
+		it("isEmptyObject({ hello: 'world' }) = false", () => {
 			expect(strings.isEmptyObject({ hello: 'world' })).to.equal(false)
 		})
 	})
 
 	describe('Test isEmptyString', () => {
-		it('isEmptyString("") = true', () => {
+		it("isEmptyString('') = true", () => {
 			expect(strings.isEmptyString('')).to.equal(true)
 		})
 
-		it('isEmptyString("hello world") = false', () => {
+		it("isEmptyString('hello world') = false", () => {
 			expect(strings.isEmptyString('hello world')).to.equal(false)
 		})
 	})
 
 	describe('Test isIncluded', () => {
-		it('isIncluded("hello world", "hello") = true', () => {
+		it("isIncluded('hello world', 'hello') = true", () => {
 			expect(strings.isIncluded('hello world', 'hello')).to.equal(true)
 		})
 
-		it('isIncluded("hello world", "earth") = false', () => {
+		it("isIncluded('hello world', 'earth') = false", () => {
 			expect(strings.isIncluded('hello world', 'earth')).to.equal(false)
 		})
 	})
@@ -234,11 +244,11 @@ describe('Test Strings Package', () => {
 	})
 
 	describe('Test isObject', () => {
-		it('isObject({ hello: "world" }) = true', () => {
+		it("isObject({ hello: 'world' }) = true", () => {
 			expect(strings.isObject({ hello: 'world' })).to.equal(true)
 		})
 
-		it('isObject("hello world") = false', () => {
+		it("isObject('hello world') = false", () => {
 			expect(strings.isObject('hello world')).to.equal(false)
 		})
 	})
@@ -254,7 +264,7 @@ describe('Test Strings Package', () => {
 	})
 
 	describe('Test notEmptyArray', () => {
-		it('notEmptyArray(["hello world"]) = true', () => {
+		it("notEmptyArray(['hello world']) = true", () => {
 			expect(strings.notEmptyArray(['hello world'])).to.equal(true)
 		})
 
@@ -264,7 +274,7 @@ describe('Test Strings Package', () => {
 	})
 
 	describe('Test notEmptyObject', () => {
-		it('notEmptyObject({ hello: "world" }) = true', () => {
+		it("notEmptyObject({ hello: 'world' }) = true", () => {
 			expect(strings.notEmptyObject({ hello: 'world' })).to.equal(true)
 		})
 
@@ -274,7 +284,7 @@ describe('Test Strings Package', () => {
 	})
 
 	describe('Test notNullOrUndefined', () => {
-		it('notNullOrUndefined("hello world") = true', () => {
+		it("notNullOrUndefined('hello world') = true", () => {
 			expect(strings.notNullOrUndefined('hello world')).to.equal(true)
 		})
 
@@ -288,29 +298,29 @@ describe('Test Strings Package', () => {
 	})
 
 	describe('Test pluralize', () => {
-		it('pluralize(1, "Apple") = 1 Apple', () => {
+		it("pluralize(1, 'Apple') = '1 Apple'", () => {
 			expect(strings.pluralize(1, 'Apple')).to.equal('1 Apple')
 		})
 
-		it('pluralize(2, "Apple") = 2 Apples', () => {
+		it("pluralize(2, 'Apple') = '2 Apples'", () => {
 			expect(strings.pluralize(2, 'Apple')).to.equal('2 Apples')
 		})
 
-		it('pluralize(1, "Child", "Children") = 1 Child', () => {
+		it("pluralize(1, 'Child', 'Children') = '1 Child'", () => {
 			expect(strings.pluralize(1, 'Child', 'Children')).to.equal('1 Child')
 		})
 
-		it('pluralize(2, "Child", "Children") = 2 Children', () => {
+		it("pluralize(2, 'Child', 'Children') = '2 Children'", () => {
 			expect(strings.pluralize(2, 'Child', 'Children')).to.equal('2 Children')
 		})
 	})
 
 	describe('Test removeDoubleSpaces', () => {
-		it('removeDoubleSpaces("hello  world")) = hello world', () => {
+		it("removeDoubleSpaces('hello  world')) = 'hello world'", () => {
 			expect(strings.removeDoubleSpaces('hello  world')).to.equal('hello world')
 		})
 
-		it('removeDoubleSpaces("hello  world  once  again")) = hello world once again', () => {
+		it("removeDoubleSpaces('hello  world  once  again')) = 'hello world once again'", () => {
 			expect(strings.removeDoubleSpaces('hello  world  once  again')).to.equal(
 				'hello world once again'
 			)
@@ -318,7 +328,7 @@ describe('Test Strings Package', () => {
 	})
 
 	describe('Test toHex', () => {
-		it('toHex("hello world")) = 68656c6c6f20776f726c64', () => {
+		it("toHex('hello world')) = '68656c6c6f20776f726c64'", () => {
 			expect(strings.toHex('hello world')).to.equal('68656c6c6f20776f726c64')
 		})
 	})